pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "money-flow-momentum"
version = "1.0.0"
description = "Production-ready Bybit V5 cross-sectional flow impulse trading bot"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Trading Bot"}]
keywords = ["bybit", "crypto", "trading", "flow", "momentum"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pybit>=5.7.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "websockets>=12.0",
    "aiohttp>=3.9.0",
    "typer>=0.9.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[project.scripts]
flow-bot = "src.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
